V6_new: Settings menus fixed/modified added virtual keyboard input

+++ app/src/main/java/com/midi_control/midi_tiles/midi/visualizer/MidiVisualizerContract.java
@@ -18,8 +18,8 @@ public interface MidiVisualizerContract {
         float getSlideSpeed(); // returns pxPerSec
 
 
-        void setMinMaxPitches(MyMath.Cords<Byte> coefficients);
-        MyMath.Cords<Byte> getMinMaxPitches();
+        void setMinMaxPitches(MyMath.Cords<Integer> coefficients);
+        MyMath.Cords<Integer> getMinMaxPitches();
         // for other
     }
